creation of unit tests

+++ internal/domain/operations.go
@@ -7,10 +7,10 @@ const (
 
 type Operation struct {
 	Operation string  `json:"operation"`
-	UnitCost  float32 `json:"unit-cost"`
+	UnitCost  float64 `json:"unit-cost"`
 	Quantity  int     `json:"quantity"`
 }
 
 type Tax struct {
-	Tax float32 `json:"tax"`
+	Tax float64 `json:"tax"`
 }
