+++ internal/utils/round.go
@@ -2,6 +2,6 @@ package utils
 
 import "math"
 
-func Round(x float32) float32 {
-	return float32(math.Round(float64(x)*100) / 100)
+func Round(x float64) float64 {
+	return math.Round(x*100) / 100
 }
